Clean up the ripper's event monitoring
We now have a function that refreshes the known state of a device based
solely on a pyudev Device object, and generates all of the necessary
events compared to the last known state. We also reuse this logic for
initialization by triggering a synthetic event for every discovered
CD-ROM device when we start the controller thread. This cleanly solves
the race between the udev monitor startup and the iteration of existing
CD-ROM devices.