Implement the most minimal possible ripping logic
This shows that we can run ddrescue with the correct arguments, and wait
for it to succeed or fail. There is no persistent state tracking at the
moment, which will be important for the sending side. There is also no
handling for a disc being ejected or replaced during the rip. However,
this is a notable waypoint to have reached.