Implement ripper state tracking and cleanup
We now clean up failed rips as soon as ddrescue exits with an error, and
have the ability to clean up failed rips on startup thanks to our state
tracking. This is also one of the big things we need for the sender: the
ability to know which ISOs are ready to be sent.