Include diskseq in DriveLoaded events
This gives us a reliable way to track disc changes even when labels are
missing or equivalent between discs.

I've added another TODO about tracking this in the database, so we can
know if we're about to re-rip a disc that's in the process of being sent
(if the isopod daemon gets restarted but allows an ongoing rip to finish
before it exits).