Introduce the "source hash" concept
Instead of tracking a separate "last rip" state, or using individual
source columns for every disc, the source hash represents a particular
combination of boot ID, device path, and diskseq as a single opaque
value. With a bit more integration work, this should be a much better
way to track previously-ripped discs.