Rework ripper controller initialization
I don't want to query the device state until the thread actually starts
running, to reduce the risk of any weird inconsistency between udev
monitoring and the initial state.

The current setup prevents losing events, but there is a possibility for
disk sequences to be delivered out of order. Maybe it's time to take
proper advantage of the whole "monotonically increasing" thing.