Filter out events with stale diskseq values
This fixes a race between loading the initial device state, starting the
udev monitor, and triggering the initial event using the original device
state. If the udev monitor comes in with a newer diskseq before we emit
the original, we don't want our later emission of the original to mess
things up.

There is still a bit of a problem with the DrivePreloaded event, which
is sort of implicitly initialized with the original diskseq... it makes
more sense in my head than in a commit message, so let me just go fix
it.