Start reworking the sender design
This is my attempt to do a more Kubernetes-style control loop design,
which I wanted to do from the start but didn't know about all the stuff
like diskseq that would help me pull it off. I'm hoping that this will
help with clean termination (i.e. the "waiting for disk space forever"
issue), and I suspect that it'll be a massive boon for status reporting.

I want to get something basic implemented for the sender first, which I
can later rework into a more generic reconciler abstraction.