Die hard on reconcile exceptions
The previous behavior was to terminate the single reconciler thread but
let the rest of the process continue. We either need to restart the
thread in those situations, or hard-kill the process like we're now
doing (which seems like the right option to me, so something like
systemd can come in and restart it with completely fresh state).