Change cleanup to shutdown
This still isn't right, though. We need to keep the udev observer alive
until the ripper process is done, since that's how we know we need to
kill it. We also need to prevent the reconciler from starting a new rip
while the controller is in a canceled state. Maybe the daemon thread
thing was useful after all?

Maybe the right approach is to spawn a new non-daemon thread from
shutdown() that listens to device events and also waits on the process?
Except then we get back to the whole "waiting on two things is a mess"
problem.

Maybe the _real_ right approach is to switch everything here to asyncio
so I can use asyncio.gather? This use case isn't actually a terrible fit
for that model.