Restart controllers with exceptions
I don't think this is the right behavior at a program level, but I think
it's the right behavior for one piece of the program, and that using
threading.excepthook on a per-program basis is a better solution for my
own needs (e.g. exceptions in the MonitorObserver thread, which still
have the same issue as those in the controller threads).