Rework the ripper arguments
Now that I'm back home, I'm testing with real hardware for the first
time. I'm not sure _exactly_ why (the paper label on the test disc is my
main suspect), but I had a ton of problems under non-direct I/O with the
drive fully resetting itself during read errors.

The change to direct I/O probably makes sense either way, but I'm not
sure if removing the retry attempts and timeout is really the right
move. This might require some back and forth with users in the field.