Add a quick read check before ddrescue
If you "accidentally" insert an audio CD into the drive, the read errors
are so pervasive that ddrescue gets stuck in an unkillable state until
you reboot the system or physically unplug the drive.

If we can't read the volume descriptor on sector 16, there's no real
hope of rescuing the disc anyways.