Add the original EPD image files
Spoiler alert: I created these in the GNU Image Manipulation Program by
creating a canvas of the exact required size (or something like that)
and playing with the icons and text until I found combinations I liked
enough. These are a few kilobytes each, so I think it's fair for them to
live in the repo, at least for now.