Refactor a lot of the EPD modules
I'm trying to set this up so that more of the internal EPD functionality
is available without the Adafruit driver modules installed. For example,
I can now run `isopod-scripts epd list` and `epd bucket` in the VM on my
laptop, rather than having to go to my actual Isopod host.