actually actually actually fix the bucket???
Am I defective as a programmer for having written such a broken
implementation of this algorithm myself instead of copying it from
elsewhere, or is thinking through the complexities of an algorithm like
this on my own supposed to be a learning experience that ultimately
makes me stronger or whatever?

Perhaps it would have been the latter if I had actually _thought_
through this at any point while writing it, rather than hacking together
the first garbage ideas that went through my head, and then immediately
hacking other garbage ideas together as soon as something went wrong.

Anyways, the token deficit idea was correct, I was just wrong about why
it was broken the first time. The deficit is when the bucket will get up
to one full token relative to _now_, not relative to the time of the
last take. Ignoring the token deficit means telling people they have to
wait the entire fill_delay for a new token even when the bucket is
partially full.

At this point, my weird test harness (yeah, no unit testing either) is
actually showing the behavior I expect, with delays that actually
correspond to when takes are allowed and without huge streams of
zero-second delays.