Remove the unnecessary max(0, X) calls
First, we prove that `fill_time - now > 0`:

  G. available < 1
  G. fill_delay > 0

  1. 0 < 1 - available
     :: 1 - available > 0
     :: tokens_missing > 0

  2. fill_delay * tokens_missing > 0

  3. now + (fill_delay * tokens_missing) > now
     :: fill_time > now

  4. fill_time - now > 0
     :: QED

Sorry, this is kind of a weird proof format, I haven't done this kind of
thing in a while and certainly not in Git commits.

Anyways, we next prove that `ready_time - now > 0`:

  G. seconds_since_take < burst_delay
     :: now - take_time < burst_delay

  1. now < burst_delay + take_time
     :: burst_delay + take_time > now

  2. burst_delay + take_time - now > 0
     :: ready_time - now > 0
     :: QED

So, there you go. I'm not going to drop this into any comments; people
reading the code should just trust that I'm right, even if they've read
my previous commit message and/or formed their opinion about me based on
the change history of this module.