Move __init__.py to daemon.py
It's time to clean up the root of the isopod package.

The weird thing about this is the force_unlink function, which a bunch
of other modules import. However, if I move that function around in the
same change, Git stops detecting the rest of the daemon module as a
rename, which means the entire blame history is lost. So, I'm going to
relocate that function in a separate commit.

(I'm sure I got this idea from Raymond Chen, but I can't find the
specific post right now.)