Add the initial versions of my systemd units
In the end, I'm somewhat convinced that I should have written a separate
daemon to fully manage the E-Ink display, including the execution of all
the connectivity and service checks that indicate whether this is a
systemfail or oob situation. I've definitely stretched the boundaries of
my understanding of systemd, and it kind of feels like I'm stretching
the boundaries of what it "wants" to do (or maybe it's that I'm still
thinking too much in terms of event handling rather than declarative
relationships?).

Again, these are just the initial versions that I put together through
trial and error. I'll try to get them better documented.