Experiment with ddrescue output in a namespaced journal
I've always been uncomfortable with the fact that the ddrescue output is
sent to /dev/null, which makes it impossible to monitor the actual
status of the rip (especially if the disc has any read errors at all).
At the same time, I don't want to manage a bunch of log files whose size
is normally far smaller than the ISOs being ripped, but could grow
larger and larger if ddrescue gets stuck retrying a bunch of errors.

This change adds experimental support for sending the output of ddrescue
to a systemd journal namespace, which can be customized to have a small
maximum size and even use volatile storage if desired. systemd usually
exposes journal namespaces using a bind mount over the standard location
of the journal socket, and journald _really_ likes to shut down when
nobody is using its socket, so I had to get a bit creative with
systemd-run and systemd-cat to get a useful file descriptor.