Implement initial support for displaying uploads
I want the E-Ink display to give some indication of how many uploads are
in flight. This is a starting point, where we just draw whatever is
available at the time of the screen refresh. Later on, I'm going to have
the sender poll the display, and I'm going to add special functionality
to the bucket so we can defer taking a token that might be more useful
to somebody else.