Don't rip discs right after a reboot
If I'm starting up the daemon for the first time on a freshly booted
system, I don't want it to immediately start ripping a pre-inserted disc
due to its source hash (which includes the boot ID) not existing. On the
other hand, if insertion of a new disc starts up a stopped daemon within
the current boot, or if a restart somehow manages to race with a disc
swap, ripping still seems like the right choice.

As part of this I've removed the special "unknown" status from the
ripper, which made the reconciliation logic a little cleaner but
required some rework of the EPD reporter. However, I think the reporter
ultimately ended up in a better state, since I had the chance to look at
it with fresh eyes.

Note that I'm dirtying up the daemon configuration a bit: the fresh boot
check directly uses the systemd-provided $RUNTIME_DIRECTORY variable,
while the variables for state and logs are passed solely through CLI
options in the unit definition. It's probably right to settle on one or
the other at some point.