Tweak some README wording
I don't want to give the impression that Isopod's lack of support for
the kinds of discs that most people probably want to rip is some kind of
moral stance rather than a largely technical decision, and my original
wording here is not NEARLY clear enough to imply that.

Audio CDs don't have normal filesystems, so ddrescue locks up trying to
read them (this is actually the source of the "quick read" check in the
ripper itself). Copy-protected DVDs need something like libdvdcss to
authenticate with the drive, otherwise the drive throws up so many read
errors that the result isn't useful (even though Isopod will try to rip
them anyways). Of course, beyond the fact that I don't need libdvdcss
for my own use case, I have to consider the fact that I'm a citizen of
the United States of America, and that I am therefore subject to the
Digital Milennium Copyright Act, and that libdvdcss is _technically_
breaking something that _technically_ counts as a copy protection
scheme, and that while there are certain legal allowances related to
this I am definitely not enough of a lawyer or Librarian of Congress to
know them all offhand, and that I probably have better things to do than
fight with hypothetical copyright trolls on this matter...

Okay, let's be more precise: it's mostly a technical decision, and
somewhat a practical decision, and if you're disappointed about Isopod's
capabilities and don't share my irrational fears of some worst-case
encounter with the U.S. legal system and/or Microsoft Corporation,
I encourage you to explore the true value of Free and Open Source
Software in the manner that works best for you, with no particular
suggestion of what this might entail.