Put current-boot-id in the state directory
Turns out systemd removes the RuntimeDirectory by default when you stop
a unit, not just on shutdown. I could figure out how to make it persist,
but I'd rather just do this, which also eliminates the dirtiness I
mentioned when I first added the $RUNTIME_DIRECTORY use (though it kind
of introduces a new dirtiness of this special global state that must be
initialized exactly once in exactly the right spot).