Allow the daemon to start without its device
On the RPi terminal hardware, I've seen systemd and udev consider
/dev/cdrom to be totally gone after a disc is ejected, which means that
if Isopod fails while the drive is empty, you can't even restart
rsyncing until someone unplugs and replugs the drive.

Except... I can't reproduce that behavior now, and this logic is a hot
mess, so I'm going to put this on an experimental branch that I can come
back to later.