Treat a small filesystem as a copy failure
This is certainly better treating it as a temporary condition, and I
think it makes more sense than treating it as an unreadable disc (which
implies that this is a property of the disc rather than the combo of the
disc and the terminal hardware).

I *kind of* thought about adding a new status to treat it as a system
failure, but I'd rather keep that reserved for the daemon being totally
down.