Make these params instead?
I'm ready to give up on the effort to get Sphinx fully working.
Apparently it *imports* the code to be documented, which means that most
things are missing when I generate the docs from macOS, which I'd like
to do instead of reworking the network settings on my VM or dumping the
output into my shared mount after every build. More importantly, I'd
rather focus on the last bits of testing before I ship this halfway
across the United States to start on its intended job.