Bump the free space polling interval to 60s
In practice, we'll spend hours in this state at a time, filling up logs
with minimally useful updates. Cutting that in half is a good start.
It'll also give us a bit more time (on average) to fill the E-Ink token
bucket, since the disc count update will get through immediately.