Release v1.0.0
Following a period of final testing, I've decided that this version of
Isopod will be the first to ship (literally, by mail) on the original
Isopod Terminal device, and to operate in a production setting with its
intended end users.

Development on Isopod that does not directly support the needs of these
end users is now frozen. No handling of TODOs, no new features. I'll
have remote access to the terminal to deploy fixes, and I'm certainly
willing to refactor and add unit tests as part of that potential work,
but without real usability feedback and real experience with 20-year-old
DVD-Rs, the paths ahead of me are too foggy to properly navigate.

(Plus, it'll be good for me personally to step back from this code base
for a little while.)