Okay, one more README thing
My intent here really wasn't to call my parents "non-technical," since
that's not accurate. The idea was something like: the device is suitable
for non-technical users, and more specifically I built it for my
parents, but there's not an inherent relationship between those two
things... but maybe this is all an excuse and I'm covering for something
I _did_ wrongly believe at one point?

Either way, this line needs to change.